Move game logic to index.js and fix linting errors
