+++ index.html
@@ -69,8 +69,9 @@
     <script type="text/javascript" src="js/domUpdates.js"></script>
     <script type="text/javascript" src="js/Player.js"></script>
     <script type="text/javascript" src="js/Game.js"></script>
-    <script type="text/javascript" src="js/Round.js"></script>
     <script type="text/javascript" src="js/Clue.js"></script>
     <script type="text/javascript" src="js/DailyDouble.js"></script>
+    <script type="text/javascript" src="js/Round.js"></script>
+    <script type="text/javascript" src="js/index.js"></script>
   </body>
 </html>
