+++ js/Clue.js
@@ -12,11 +12,6 @@ class Clue {
   }
 }
 
-
-
-
-
-
 if (typeof module !== 'undefined') {
   module.exports = Clue;
 }
