+++ js/Game.js
@@ -1,8 +1,8 @@
 class Game {
-  constructor(round = 1, activePlayer, cluesRemaining = 16) {
-    this.round = round;
+  constructor(activePlayer) {
+    this.round = 1;
     this.activePlayer = activePlayer;
-    this.cluesRemaining = cluesRemaining;
+    this.cluesRemaining = 16;
   }
 
   reset() {
@@ -13,7 +13,6 @@ class Game {
     
   }
 }
-
 
 if (typeof module !== 'undefined') {
   module.exports = Game;
