+++ js/Round.js
@@ -41,7 +41,6 @@ class Round {
   }
 }
 
-
 if (typeof module !== 'undefined') {
   module.exports = Round;
 }
