+++ js/dataset.js
@@ -706,7 +706,3 @@ const data = {
     }
   ]
 };
-
-if (typeof module !== 'undefined') {
-  module.exports = data;
-}
