+++ test/DailyDouble-test.js
@@ -2,16 +2,15 @@ const chai = require('chai');
 const expect = chai.expect;
 const DailyDouble = require('../js/DailyDouble.js');
 
-
 describe('DailyDouble', function() {
   it('should instantiate a new DailyDouble', function() {
-    const dailyDouble = new DailyDouble('q', 'a', 100, 1)
-    expect(dailyDouble.question).to.equal('q')
-    expect(dailyDouble.answer).to.equal('a')
-    expect(dailyDouble.pointValue).to.equal(100)
-    expect(dailyDouble.categoryId).to.equal(1)
-    expect(dailyDouble.dailyDouble).to.equal(true)
-  })
+    const dailyDouble = new DailyDouble('q', 'a', 100, 1);
+    expect(dailyDouble.question).to.equal('q');
+    expect(dailyDouble.answer).to.equal('a');
+    expect(dailyDouble.pointValue).to.equal(100);
+    expect(dailyDouble.categoryId).to.equal(1);
+    expect(dailyDouble.dailyDouble).to.equal(true);
+  });
 });
